chore: 배포 환경 이전

+++ src/main/resources/db/migration/V3__add_class_no_to_code.sql
@@ -1,4 +1,4 @@
-ALTER TABLE code ADD COLUMN class_no INT;
+ALTER TABLE code ADD COLUMN class_no BIGINT;
 
 UPDATE code
 SET class_no = u.class_number
